Create ImageResizer
This is my playing around with basic openCV. This program will take a file path that has images in it from the user and spit out a resized copy of it. Either in back and white or colour. Still need to address the fact that the user has to put r in front of their file path. i.e r"C:\Users\user\Desktop\images".